refactor: GameSyncService 클래스 분리
