+++ backend/src/main/java/com/yagubogu/game/schedule/GameScheduler.java
@@ -1,6 +1,7 @@
 package com.yagubogu.game.schedule;
 
 import com.yagubogu.game.exception.GameSyncException;
+import com.yagubogu.game.service.GameResultSyncService;
 import com.yagubogu.game.service.GameSyncService;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +13,7 @@ import org.springframework.stereotype.Component;
 public class GameScheduler {
 
     private final GameSyncService gameSyncService;
+    private final GameResultSyncService gameResultSyncService;
 
     @Scheduled(cron = "0 0 0 * * *")
     public void fetchDailyGameSchedule() {
@@ -27,7 +29,7 @@ public class GameScheduler {
     public void fetchDailyGameResult() {
         LocalDate yesterday = LocalDate.now().minusDays(1);
         try {
-            gameSyncService.syncGameResult(yesterday);
+            gameResultSyncService.syncGameResult(yesterday);
         } catch (GameSyncException e) {
             // TODO: 예외 로깅
         }
