+++ backend/src/main/java/com/yagubogu/game/service/GameResultSyncService.java
@@ -1,0 +1,47 @@
+package com.yagubogu.game.service;
+
+import com.yagubogu.game.domain.Game;
+import com.yagubogu.game.domain.ScoreBoard;
+import com.yagubogu.game.dto.KboGameResponse;
+import com.yagubogu.game.dto.KboGameResultResponse;
+import com.yagubogu.game.repository.GameRepository;
+import com.yagubogu.game.service.client.KboClient;
+import java.time.LocalDate;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class GameResultSyncService {
+
+    private final KboClient kboClient;
+    private final GameRepository gameRepository;
+
+    @Transactional
+    public void syncGameResult(LocalDate date) {
+        List<KboGameResponse> gameResponses = kboClient.fetchGames(date).games();
+
+        for (KboGameResponse response : gameResponses) {
+            gameRepository.findByGameCode(response.gameCode())
+                    .ifPresent(game -> updateGameDetails(game, response));
+        }
+    }
+
+    private void updateGameDetails(Game game, KboGameResponse response) {
+        game.updateGameState(response.gameState());
+
+        if (game.getGameState().isNotCompleted()) {
+            // TODO: gameState가 LIVE 일 때 로깅
+            return;
+        }
+
+        KboGameResultResponse gameResult = kboClient.fetchGameResult(game);
+        ScoreBoard homeScoreBoard = gameResult.homeScoreBoard().toScoreBoard();
+        ScoreBoard awayScoreBoard = gameResult.awayScoreBoard().toScoreBoard();
+
+        game.updateScoreBoard(homeScoreBoard, awayScoreBoard);
+    }
+}
